1294: Rename verification package to cards
