+++ administration/src/graphql/cards/createCards.graphql
