+++ administration/src/graphql/cards/deleteInactiveCards.graphql
