+++ administration/src/graphql/cards/getCardStatisticsInProject.graphql
