+++ administration/src/graphql/cards/getCardStatisticsInRegion.graphql
