+++ administration/src/graphql/cards/sendCardConfirmationMail.graphql
