+++ backend/src/main/kotlin/app/ehrenamtskarte/backend/cards/service/CardActivator.kt
@@ -1,6 +1,6 @@
-package app.ehrenamtskarte.backend.verification.service
+package app.ehrenamtskarte.backend.cards.service
 
-import app.ehrenamtskarte.backend.verification.database.TOTP_SECRET_LENGTH
+import app.ehrenamtskarte.backend.cards.database.TOTP_SECRET_LENGTH
 import at.favre.lib.crypto.bcrypt.BCrypt
 import com.eatthepath.otp.TimeBasedOneTimePasswordGenerator
 import javax.crypto.KeyGenerator
