+++ backend/src/main/kotlin/app/ehrenamtskarte/backend/cards/service/CardVerifier.kt
@@ -1,9 +1,9 @@
-package app.ehrenamtskarte.backend.verification.service
+package app.ehrenamtskarte.backend.cards.service
 
-import app.ehrenamtskarte.backend.verification.ValidityPeriodUtil.Companion.daysSinceEpochToDate
-import app.ehrenamtskarte.backend.verification.ValidityPeriodUtil.Companion.isOnOrAfterToday
-import app.ehrenamtskarte.backend.verification.ValidityPeriodUtil.Companion.isOnOrBeforeToday
-import app.ehrenamtskarte.backend.verification.database.repos.CardRepository
+import app.ehrenamtskarte.backend.cards.ValidityPeriodUtil.Companion.daysSinceEpochToDate
+import app.ehrenamtskarte.backend.cards.ValidityPeriodUtil.Companion.isOnOrAfterToday
+import app.ehrenamtskarte.backend.cards.ValidityPeriodUtil.Companion.isOnOrBeforeToday
+import app.ehrenamtskarte.backend.cards.database.repos.CardRepository
 import com.eatthepath.otp.TimeBasedOneTimePasswordGenerator
 import org.jetbrains.exposed.sql.transactions.transaction
 import java.time.Duration
