+++ backend/src/main/kotlin/app/ehrenamtskarte/backend/cards/webservice/schema/CardMutationService.kt
@@ -1,9 +1,21 @@
-package app.ehrenamtskarte.backend.verification.webservice.schema
+package app.ehrenamtskarte.backend.cards.webservice.schema
 
 import Card
 import app.ehrenamtskarte.backend.application.database.repos.ApplicationRepository
 import app.ehrenamtskarte.backend.auth.database.AdministratorEntity
 import app.ehrenamtskarte.backend.auth.service.Authorizer
+import app.ehrenamtskarte.backend.cards.PEPPER_LENGTH
+import app.ehrenamtskarte.backend.cards.database.CodeType
+import app.ehrenamtskarte.backend.cards.database.repos.CardRepository
+import app.ehrenamtskarte.backend.cards.hash
+import app.ehrenamtskarte.backend.cards.service.CardActivator
+import app.ehrenamtskarte.backend.cards.service.CardVerifier
+import app.ehrenamtskarte.backend.cards.webservice.QRCodeUtil
+import app.ehrenamtskarte.backend.cards.webservice.schema.types.ActivationState
+import app.ehrenamtskarte.backend.cards.webservice.schema.types.CardActivationResultModel
+import app.ehrenamtskarte.backend.cards.webservice.schema.types.CardCreationResultModel
+import app.ehrenamtskarte.backend.cards.webservice.schema.types.DynamicActivationCodeResult
+import app.ehrenamtskarte.backend.cards.webservice.schema.types.StaticVerificationCodeResult
 import app.ehrenamtskarte.backend.common.webservice.GraphQLContext
 import app.ehrenamtskarte.backend.exception.service.ForbiddenException
 import app.ehrenamtskarte.backend.exception.service.ProjectNotFoundException
@@ -15,18 +27,6 @@ import app.ehrenamtskarte.backend.exception.webservice.exceptions.RegionNotFound
 import app.ehrenamtskarte.backend.mail.Mailer
 import app.ehrenamtskarte.backend.matomo.Matomo
 import app.ehrenamtskarte.backend.regions.database.repos.RegionsRepository
-import app.ehrenamtskarte.backend.verification.PEPPER_LENGTH
-import app.ehrenamtskarte.backend.verification.database.CodeType
-import app.ehrenamtskarte.backend.verification.database.repos.CardRepository
-import app.ehrenamtskarte.backend.verification.hash
-import app.ehrenamtskarte.backend.verification.service.CardActivator
-import app.ehrenamtskarte.backend.verification.service.CardVerifier
-import app.ehrenamtskarte.backend.verification.webservice.QRCodeUtil
-import app.ehrenamtskarte.backend.verification.webservice.schema.types.ActivationState
-import app.ehrenamtskarte.backend.verification.webservice.schema.types.CardActivationResultModel
-import app.ehrenamtskarte.backend.verification.webservice.schema.types.CardCreationResultModel
-import app.ehrenamtskarte.backend.verification.webservice.schema.types.DynamicActivationCodeResult
-import app.ehrenamtskarte.backend.verification.webservice.schema.types.StaticVerificationCodeResult
 import com.expediagroup.graphql.generator.annotations.GraphQLDescription
 import com.google.protobuf.ByteString
 import extensionStartDayOrNull
