+++ backend/src/main/kotlin/app/ehrenamtskarte/backend/cards/webservice/schema/types/CardVerificationResultModel.kt
@@ -1,4 +1,4 @@
-package app.ehrenamtskarte.backend.verification.webservice.schema.types
+package app.ehrenamtskarte.backend.cards.webservice.schema.types
 
 import java.time.Instant
