+++ backend/src/main/kotlin/app/ehrenamtskarte/backend/exception/webservice/exceptions/InvalidQrCodeSize.kt
@@ -1,7 +1,7 @@
 package app.ehrenamtskarte.backend.exception.webservice.exceptions
 
+import app.ehrenamtskarte.backend.cards.database.CodeType
 import app.ehrenamtskarte.backend.exception.webservice.schema.GraphQLExceptionCode
-import app.ehrenamtskarte.backend.verification.database.CodeType
 import graphql.GraphqlErrorException
 
 class InvalidQrCodeSize(encodedCardInfoBase64: String, codeType: CodeType) : GraphqlErrorException(
