+++ backend/src/main/kotlin/app/ehrenamtskarte/backend/migration/TablesRegistry.kt
@@ -3,6 +3,7 @@ package app.ehrenamtskarte.backend.migration
 import app.ehrenamtskarte.backend.application.database.ApplicationVerifications
 import app.ehrenamtskarte.backend.application.database.Applications
 import app.ehrenamtskarte.backend.auth.database.Administrators
+import app.ehrenamtskarte.backend.cards.database.Cards
 import app.ehrenamtskarte.backend.migration.database.Migrations
 import app.ehrenamtskarte.backend.projects.database.Projects
 import app.ehrenamtskarte.backend.regions.database.Regions
@@ -11,7 +12,6 @@ import app.ehrenamtskarte.backend.stores.database.Addresses
 import app.ehrenamtskarte.backend.stores.database.Categories
 import app.ehrenamtskarte.backend.stores.database.Contacts
 import app.ehrenamtskarte.backend.stores.database.PhysicalStores
-import app.ehrenamtskarte.backend.verification.database.Cards
 
 object TablesRegistry {
     fun getAllTables() = arrayOf(
