+++ backend/src/test/kotlin/app/ehrenamtskarte/backend/cards/CanonicalJsonTest.kt
@@ -1,4 +1,4 @@
-package app.ehrenamtskarte.backend.verification
+package app.ehrenamtskarte.backend.cards
 
 import Card
 import app.ehrenamtskarte.backend.helper.CardInfoTestSample
