+++ frontend/graphql_queries/cards/card_activation.graphql
