+++ frontend/graphql_queries/cards/card_verification_by_hash.graphql
